♻️ 39YSZ1 routing: keep route modules within size budget
