+++ packages/agentplane/src/cli/run-cli.core.route-decision.test.ts
@@ -11,7 +11,6 @@ import {
   expect,
   it,
   mkGitRepoRootWithBranch,
-  recordVerificationOk,
   runCli,
   runCliSilent,
   writeConfig,
@@ -634,193 +633,6 @@ describe("runCli route decision commands", () => {
       "utf8",
     );
     expect(readme).toContain('plan_approval:\n  state: "pending"');
-  });
-
-  it("requires remote truth before declaring a done branch_pr task terminal", async () => {
-    const root = await mkGitRepoRootWithBranch("main");
-    const config = defaultConfig();
-    config.workflow_mode = "branch_pr";
-    await writeConfig(root, config);
-    await runCliSilent(["branch", "base", "set", "main", "--root", root]);
-
-    const taskId = await createBranchPrTask(root);
-    await runCliSilent([
-      "task",
-      "plan",
-      "set",
-      taskId,
-      "--text",
-      "Exercise done route decisions.",
-      "--updated-by",
-      "ORCHESTRATOR",
-      "--root",
-      root,
-    ]);
-    await runCliSilent(["task", "plan", "approve", taskId, "--by", "ORCHESTRATOR", "--root", root]);
-    await runCliSilent(["task", "set-status", taskId, "DOING", "--force", "--yes", "--root", root]);
-    await execFileAsync("git", ["add", "--all"], { cwd: root });
-    await execFileAsync("git", ["commit", "-m", "task: prepare done route fixture"], {
-      cwd: root,
-    });
-    await recordVerificationOk(root, taskId);
-
-    const readmePath = path.join(root, ".agentplane", "tasks", taskId, "README.md");
-    const readme = await readFile(readmePath, "utf8");
-    const { stdout: implementationHead } = await execFileAsync("git", ["rev-parse", "HEAD"], {
-      cwd: root,
-    });
-    await writeFile(
-      readmePath,
-      readme
-        .replace('status: "DOING"', 'status: "DONE"')
-        .replace(
-          "commit: null",
-          `commit:\n  hash: "${implementationHead.trim()}"\n  message: "Merge PR #1"`,
-        ),
-      "utf8",
-    );
-    const statusIo = captureStdIO();
-    try {
-      const code = await runCli(["task", "status", taskId, "--route", "--json", "--root", root]);
-      if (code !== 0) process.stderr.write(statusIo.stderr);
-      expect(code).toBe(0);
-      const parsed = JSON.parse(statusIo.stdout) as {
-        blockers: { code: string }[];
-        nextAction: { code: string; command: string | null };
-        oracle: { phase: string; nextCommand: string };
-      };
-      expect(parsed.blockers).toEqual([]);
-      expect(parsed.oracle.phase).toBe("side_effect_authority_required");
-      await runCliSilent([...parsed.oracle.nextCommand.split(" ").slice(1), "--root", root]);
-    } finally {
-      statusIo.restore();
-    }
-
-    const [code, parsed] = await runJson<{
-      blockers: { code: string }[];
-      nextAction: { code: string; command: string | null };
-      oracle: { phase: string };
-    }>(["task", "status", taskId, "--route", "--json", "--root", root]);
-    expect(code).toBe(0);
-    expect(parsed.blockers).toEqual([]);
-    expect(parsed.oracle.phase).toBe("remote_route_refresh_needed");
-    expect(parsed.nextAction).toMatchObject({
-      code: "refresh_remote_route",
-      command: `agentplane task next-action ${taskId} --remote --explain`,
-    });
-  });
-
-  it("invalidates a passing verification record after a semantic branch commit", async () => {
-    const root = await mkGitRepoRootWithBranch("main");
-    const config = defaultConfig();
-    config.workflow_mode = "branch_pr";
-    await writeConfig(root, config);
-    await runCliSilent(["branch", "base", "set", "main", "--root", root]);
-
-    const taskId = await createBranchPrTask(root);
-    await runCliSilent([
-      "task",
-      "plan",
-      "set",
-      taskId,
-      "--text",
-      "Require verification evidence for the current semantic implementation head.",
-      "--updated-by",
-      "PLANNER",
-      "--root",
-      root,
-    ]);
-    await runCliSilent(["task", "plan", "approve", taskId, "--by", "ORCHESTRATOR", "--root", root]);
-
-    const branch = `task/${taskId}/verification-freshness`;
-    await execFileAsync("git", ["checkout", "-b", branch], { cwd: root });
-    await runCliSilent([
-      "task",
-      "start-ready",
-      taskId,
-      "--author",
-      "CODER",
-      "--body",
-      "Start: exercise verification freshness on the task branch.",
-      "--root",
-      root,
-    ]);
-    await execFileAsync("git", ["add", "--all"], { cwd: root });
-    await execFileAsync("git", ["commit", "-m", "task: prepare verification fixture"], {
-      cwd: root,
-    });
-
-    await writeFile(path.join(root, "impl.txt"), "first implementation\n", "utf8");
-    await execFileAsync("git", ["add", "impl.txt"], { cwd: root });
-    await execFileAsync("git", ["commit", "-m", "feat: first implementation"], { cwd: root });
-    const { stdout: firstImplementationHead } = await execFileAsync("git", ["rev-parse", "HEAD"], {
-      cwd: root,
-    });
-    await runCliSilent([
-      "task",
-      "set-status",
-      taskId,
-      "DOING",
-      "--commit",
-      firstImplementationHead.trim(),
-      "--author",
-      "CODER",
-      "--body",
-      "Implementation committed: first semantic implementation.",
-      "--root",
-      root,
-    ]);
-    await recordVerificationOk(root, taskId);
-    await execFileAsync("git", ["add", "--all"], { cwd: root });
-    await execFileAsync("git", ["commit", "-m", "task: record fresh verification"], {
-      cwd: root,
-    });
-
-    const [, fresh] = await runJson<{ blockers: { code: string; summary: string }[] }>([
-      "task",
-      "status",
-      taskId,
-      "--route",
-      "--json",
-      "--root",
-      root,
-    ]);
-    expect(fresh.blockers.map((blocker) => blocker.code)).not.toContain("verification_required");
-
-    await writeFile(path.join(root, "impl.txt"), "second implementation\n", "utf8");
-    await execFileAsync("git", ["add", "impl.txt"], { cwd: root });
-    await execFileAsync("git", ["commit", "-m", "feat: second implementation"], { cwd: root });
-    const { stdout: secondImplementationHead } = await execFileAsync("git", ["rev-parse", "HEAD"], {
-      cwd: root,
-    });
-    await runCliSilent([
-      "task",
-      "set-status",
-      taskId,
-      "DOING",
-      "--commit",
-      secondImplementationHead.trim(),
-      "--author",
-      "CODER",
-      "--body",
-      "Implementation committed: second semantic implementation.",
-      "--root",
-      root,
-    ]);
-
-    const [, stale] = await runJson<{ blockers: { code: string; summary: string }[] }>([
-      "task",
-      "status",
-      taskId,
-      "--route",
-      "--json",
-      "--root",
-      root,
-    ]);
-    expect(stale.blockers).toContainEqual({
-      code: "verification_required",
-      summary: "the passing verification record does not cover the current implementation head",
-    });
   });
 
   it("routes stale local branch_pr state to base sync after hosted close is recorded upstream", async () => {
